+++ app/pom.xml
@@ -43,6 +43,11 @@
             <artifactId>apicurio-registry-tenant-manager-client</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>io.apicurio</groupId>
+            <artifactId>apicurio-registry-utils-import-export</artifactId>
+        </dependency>
+
         <!-- Quarkus Dependencies -->
         <dependency>
             <groupId>io.quarkus</groupId>
