+++ app/src/main/java/io/apicurio/registry/storage/impexp/EntityInputStream.java
@@ -19,11 +19,13 @@ package io.apicurio.registry.storage.impexp;
 import java.io.Closeable;
 import java.io.IOException;
 
+import io.apicurio.registry.utils.impexp.Entity;
+
 /**
  * @author eric.wittmann@gmail.com
  */
 public interface EntityInputStream extends Closeable {
-    
+
     /**
      * Get the next import entity from the stream of entities being imported.
      */
