+++ app/src/main/java/io/apicurio/registry/storage/impl/sql/mappers/ArtifactRuleEntityMapper.java
@@ -22,9 +22,9 @@ import java.sql.SQLException;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.statement.StatementContext;
 
-import io.apicurio.registry.storage.impexp.ArtifactRuleEntity;
 import io.apicurio.registry.storage.impl.sql.SqlUtil;
 import io.apicurio.registry.types.RuleType;
+import io.apicurio.registry.utils.impexp.ArtifactRuleEntity;
 
 /**
  * @author eric.wittmann@gmail.com
