+++ app/src/main/java/io/apicurio/registry/storage/impl/sql/mappers/GroupEntityMapper.java
@@ -22,9 +22,9 @@ import java.sql.SQLException;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.statement.StatementContext;
 
-import io.apicurio.registry.storage.impexp.GroupEntity;
 import io.apicurio.registry.storage.impl.sql.SqlUtil;
 import io.apicurio.registry.types.ArtifactType;
+import io.apicurio.registry.utils.impexp.GroupEntity;
 
 /**
  * @author eric.wittmann@gmail.com
