+++ pom.xml
@@ -98,6 +98,8 @@
         <module>utils/kafka</module>
         <module>utils/maven-plugin</module>
         <module>utils/tools</module>
+        <module>utils/importexport</module>
+        <module>utils/exportV1</module>
         <module>schema-compatibility/json</module>
         <module>schema-compatibility/protobuf</module>
         <module>utils/tests</module>
@@ -330,6 +332,11 @@
             <dependency>
                 <groupId>io.apicurio</groupId>
                 <artifactId>apicurio-registry-tenant-manager-client</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.apicurio</groupId>
+                <artifactId>apicurio-registry-utils-import-export</artifactId>
                 <version>${project.version}</version>
             </dependency>
