+++ utils/exportV1/pom.xml
@@ -35,6 +35,11 @@
 
         <dependency>
             <groupId>io.apicurio</groupId>
+            <artifactId>apicurio-registry-utils-import-export</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>io.apicurio</groupId>
             <artifactId>apicurio-registry-rest-client</artifactId>
             <version>1.3.2.Final</version>
         </dependency>
