--- utils/exportV1/src/main/java/io/apicurio/registry/utils/export/ContentEntity.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2021 Red Hat
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.apicurio.registry.utils.export;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
-/**
- * @author eric.wittmann@gmail.com
- */
-public class ContentEntity extends Entity {
-
-    public long contentId;
-    public String canonicalHash;
-    public String contentHash;
-    @JsonIgnore
-    public byte[] contentBytes;
-
-    /**
-     * @see io.apicurio.registry.storage.impexp.Entity#getEntityType()
-     */
-    @Override
-    public EntityType getEntityType() {
-        return EntityType.Content;
-    }
-
-}
