+++ utils/importexport/src/main/java/io/apicurio/registry/utils/impexp/EntityReader.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.apicurio.registry.rest.v2.impexp;
+package io.apicurio.registry.utils.impexp;
 
 import java.io.IOException;
 import java.util.zip.ZipEntry;
@@ -24,14 +24,6 @@ import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import io.apicurio.registry.storage.impexp.ArtifactRuleEntity;
-import io.apicurio.registry.storage.impexp.ArtifactVersionEntity;
-import io.apicurio.registry.storage.impexp.ContentEntity;
-import io.apicurio.registry.storage.impexp.Entity;
-import io.apicurio.registry.storage.impexp.EntityType;
-import io.apicurio.registry.storage.impexp.GlobalRuleEntity;
-import io.apicurio.registry.storage.impexp.GroupEntity;
-import io.apicurio.registry.storage.impexp.ManifestEntity;
 import io.apicurio.registry.utils.IoUtil;
 
 /**
