+++ utils/importexport/src/main/java/io/apicurio/registry/utils/impexp/GlobalRuleEntity.java
@@ -14,25 +14,24 @@
  * limitations under the License.
  */
 
-package io.apicurio.registry.storage.impexp;
+package io.apicurio.registry.utils.impexp;
 
 import io.apicurio.registry.types.RuleType;
 
 /**
  * @author eric.wittmann@gmail.com
  */
-public class ArtifactRuleEntity extends Entity {
+public class GlobalRuleEntity extends Entity {
 
-    public String groupId;
-    public String artifactId;
-    public RuleType type;
+    public RuleType ruleType;
     public String configuration;
 
     /**
-     * @see io.apicurio.registry.storage.impexp.Entity#getEntityType()
+     * @see io.apicurio.registry.utils.impexp.storage.impexp.Entity#getEntityType()
      */
     @Override
     public EntityType getEntityType() {
-        return EntityType.ArtifactRule;
+        return EntityType.GlobalRule;
     }
+
 }
