+++ utils/importexport/src/main/java/io/apicurio/registry/utils/impexp/GroupEntity.java
@@ -14,41 +14,32 @@
  * limitations under the License.
  */
 
-package io.apicurio.registry.utils.export;
+package io.apicurio.registry.utils.impexp;
 
-import java.util.List;
 import java.util.Map;
 
-import io.apicurio.registry.types.ArtifactState;
 import io.apicurio.registry.types.ArtifactType;
 
 /**
  * @author eric.wittmann@gmail.com
  */
-public class ArtifactVersionEntity extends Entity {
+public class GroupEntity extends Entity {
 
-    public long globalId;
     public String groupId;
-    public String artifactId;
-    public String version;
-    public int versionId;
-    public ArtifactType artifactType;
-    public ArtifactState state;
-    public String name;
     public String description;
+    public ArtifactType artifactsType;
     public String createdBy;
     public long createdOn;
-    public List<String> labels;
+    public String modifiedBy;
+    public long modifiedOn;
     public Map<String, String> properties;
-    public boolean isLatest;
-    public long contentId;
 
     /**
-     * @see io.apicurio.registry.storage.impexp.Entity#getEntityType()
+     * @see io.apicurio.registry.utils.impexp.storage.impexp.Entity#getEntityType()
      */
     @Override
     public EntityType getEntityType() {
-        return EntityType.ArtifactVersion;
+        return EntityType.Group;
     }
 
 }
